+++ AtlassianPS.Configuration/AtlassianPS.Configuration.psm1
@@ -18,6 +18,7 @@ if (-not("AtlassianPS.ServerData" -as [Type])) {
 if ($PSVersionTable.PSVersion.Major -lt 5) {
     Add-Type -Path (Join-Path $PSScriptRoot AtlassianPS.Configuration.Attributes.cs) -ReferencedAssemblies Microsoft.CSharp, Microsoft.PowerShell.Commands.Utility, System.Management.Automation, System.Runtime.Extensions, System.Security.Cryptography.X509Certificates
 }
+
 #endregion Dependencies
 
 #region ModuleConfig
@@ -37,8 +38,14 @@ Add-MetadataConverter @{
 }
 
 #region LoadFunctions
-$PublicFunctions = @( Get-ChildItem -Path "$PSScriptRoot/Public/*.ps1" -ErrorAction SilentlyContinue )
-$PrivateFunctions = @( Get-ChildItem -Path "$PSScriptRoot/Private/*.ps1" -ErrorAction SilentlyContinue )
+$PublicFunctions = @(
+    Get-ChildItem -Path "$PSScriptRoot/Public" -Recurse -File -Filter "*.ps1" -ErrorAction SilentlyContinue |
+        Sort-Object -Property FullName
+)
+$PrivateFunctions = @(
+    Get-ChildItem -Path "$PSScriptRoot/Private" -Recurse -File -Filter "*.ps1" -ErrorAction SilentlyContinue |
+        Sort-Object -Property FullName
+)
 
 # Dot source the functions
 foreach ($file in @($PublicFunctions + $PrivateFunctions)) {